(#2709) - `revpos` property on attachment create
Add the `revpos` property to the attachment stub. The `revpos` is the doc's
`rev` prefix (a sequential number) from when the attachment was added or
updated. It is used during replication in the CouchDB replicator (and others
including CouchBase Lite) to avoid copying attachments that have not changed
since the last replication.

This solves a problem with documents with attachments created in PouchDB, which
are later replicated to an Apache CouchDB. If the doc is then pulled from an iOS
application using CouchBase Lite the doc is skipped, because CBL rejects the
attachments because of invalid attachment stub (see
https://github.com/couchbase/couchbase-lite-ios/issues/1200)

Also this improves replication performance of documents created by PouchDB,
which were later synced using Apache CouchDB replicator.